refactor config parser code to avoid daemon package in cli
